projeto2/requirements copy.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-bootstrap4==3.0.1
django-stdimage==5.3.0
flake8==3.9.2
gunicorn==20.1.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
Pillow==8.3.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyMySQL==1.0.2
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
whitenoise==5.3.0
